backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
ortools==9.8.3296
firebase-admin==6.4.0